requirements.txt
Flask==2.3.2
natsort==8.2.0
numpy==1.23.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==9.4.0
tqdm==4.64.1
